+++ src/compas_ghpython/utilities/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
+from warnings import warn
 
-from .drawing import (
+from ..drawing import (
     draw_frame,
     draw_points,
     draw_lines,
@@ -31,3 +32,5 @@ __all__ = [
     "draw_circles",
     "draw_brep",
 ]
+
+warn("compas_ghpython.utilities will be removed in version 2.3. Please use compas_ghpython.drawing instead.", DeprecationWarning, stacklevel=2)
